pyproject.toml
[project]
name = "yt-elt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest==8.3.3",
    "python-dotenv>=1.2.1",
    "soda-core-postgres==3.3.14",
]
